See #39. Moves configuration into the util file
