+++ src/pi-src/util.h
@@ -6,5 +6,7 @@ Project: Crab Tracker
 Created: 2018-02-19
 ******************************************************************************/
 
+int initialize_util();
+int get_param(char*, int*);
 void print_bin_8(uint8_t);
 void print_bin_ulong(unsigned long);
